+++ src/command_modules/azure-cli-botservice/HISTORY.rst
@@ -2,6 +2,9 @@
 
 Release History
 ===============
+
+1.0.0
++++++
 * [Breaking Change]: The default value for `--version` is now `v4`, not `v3` (except for `az bot prepare-publish`).
 * [Breaking Change]: `--lang` no longer defaults to `Csharp`. If the command requires `--lang` and it is not provided, the command will error out.
 * [Breaking Change]: The `--appid` and `--password` args for `az bot create` are now required and can be created via `az ad app create`.